finish splitting index.ts
